+++ src/controls/Search.js
@@ -36,7 +36,10 @@ class SearchControl {
                         boundingbox,
                     } = result
                     const [y1, y2, x1, x2] = boundingbox
-                    const bounds = [[x1, y1], [x2, y2]]
+                    const bounds = [
+                        [x1, y1],
+                        [x2, y2],
+                    ]
 
                     return { name, lat, lng, bounds }
                 })