+++ src/utils/earthengine.js
@@ -1,9 +1,12 @@
 export const defaultEarthEngineOptions = {
     bandReducer: 'sum',
     popup: '{name}: {value} {unit}',
+    nullPopup: '{name}: {noValue}',
+    noValue: 'no value',
 }
 
 const earthEngineOptions = [
+    'format',
     'aggregationType',
     'band',
     'bandReducer',
@@ -19,10 +22,19 @@ const earthEngineOptions = [
 ]
 
 // Returns the layer options that should be passed to the EE worker
-export const getEarthEngineOptions = options =>
-    Object.keys(options)
+export const getEarthEngineOptions = opts => {
+    console.log('hey getEarthEngineOptions', opts)
+    const options = Object.keys(opts)
         .filter(option => earthEngineOptions.includes(option))
         .reduce((obj, key) => {
-            obj[key] = options[key]
+            obj[key] = opts[key]
             return obj
         }, {})
+
+    // Exclude point features if no buffer
+    if (options.data && !options.buffer) {
+        options.data = options.data.filter(d => d.geometry.type !== 'Point')
+    }
+
+    return options
+}